+++ lib/Controller/ConfigController.php
@@ -10,6 +10,7 @@ use OCP\AppFramework\Controller;
 use OCA\Cookbook\Service\RecipeService;
 use OCA\Cookbook\Service\DbCacheService;
 use OCA\Cookbook\Helper\RestParameterParser;
+use OCA\Cookbook\Helper\UserFolderHelper;
 
 class ConfigController extends Controller {
 	/**
@@ -27,12 +28,25 @@ class ConfigController extends Controller {
 	 */
 	private $restParser;
 
-	public function __construct($AppName, IRequest $request, RecipeService $recipeService, DbCacheService $dbCacheService, RestParameterParser $restParser) {
+	/**
+	 * @var UserFolderHelper
+	 */
+	private $userFolder;
+
+	public function __construct(
+		$AppName,
+		IRequest $request,
+		RecipeService $recipeService,
+		DbCacheService $dbCacheService,
+		RestParameterParser $restParser,
+		UserFolderHelper $userFolder
+	) {
 		parent::__construct($AppName, $request);
 
 		$this->service = $recipeService;
 		$this->dbCacheService = $dbCacheService;
 		$this->restParser = $restParser;
+		$this->userFolder = $userFolder;
 	}
 
 	/**
@@ -43,7 +57,7 @@ class ConfigController extends Controller {
 		$this->dbCacheService->triggerCheck();
 		
 		return new DataResponse([
-			'folder' => $this->service->getUserFolderPath(),
+			'folder' => $this->userFolder->getPath(),
 			'update_interval' => $this->dbCacheService->getSearchIndexUpdateInterval(),
 			'print_image' => $this->service->getPrintImage(),
 		], Http::STATUS_OK);
@@ -57,7 +71,7 @@ class ConfigController extends Controller {
 		$data = $this->restParser->getParameters();
 		
 		if (isset($data['folder'])) {
-			$this->service->setUserFolderPath($data['folder']);
+			$this->userFolder->setPath($data['folder']);
 			$this->dbCacheService->updateCache();
 		}