+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### Added
 - Create service class for downloading and extracting JSON
   [#553](https://github.com/nextcloud/cookbook/pull/553) @christianlupus
+
+### Changed
+- Extracted user folder handling into its own helper class
+  [#1007](https://github.com/nextcloud/cookbook/pull/1007) @christianlupus
   
 ### Fixed
 - Fix visual regression in edit mode to prevent overflow of breadcrumbs